Move wgpu-canvas
