+++ ffi-run/src/platform/ios.rs
@@ -1,6 +1,6 @@
 use app_surface::{AppSurface, IOSViewObj};
 use wgpu::{CurrentSurfaceTexture, Device, Queue, SurfaceConfiguration, SurfaceTexture, Texture, TextureView};
-use wgpu_canvas::wgpu_canvas::WgpuCanvas;
+use renderer::wgpu_canvas::WgpuCanvas;
 use crate::ShashlikMapApi;
 use map::tiles::default_tiles_provider::DefaultTilesProvider;
 use osm::source::reqwest_source::ReqwestSource;
