+++ wgpu-canvas/src/lib.rs
@@ -15,7 +15,6 @@ pub mod render_group;
 pub mod render_modifier;
 pub mod render_style;
 pub mod style_id;
-pub mod wgpu_canvas;
 
 /// should be the same as mesh_shader.wgsl
 pub static LIGHT_POS: DVec3 = dvec3(0.84, 1.12, 1.42);
